==> analysis_beginning/00.check_data-Copy2.ipynb <==
import pandas as pd
import numpy as np
import anndata as ad

# Metadata

!ls ../data_dump/candidate_genes*

f = '../data_dump/candidate_genes_vincent_0503.csv'
# df0 = pd.read_excel(f, sheet_name=0)
df0 = pd.read_csv(f)
print(df0.shape)

df0 = df0.groupby('Gene').first().reset_index()
print(df0.shape)
df0 = df0[~df0['Gene'].isnull()]
print(df0.shape)
df0['Gene'] = df0['Gene'].apply(lambda x: x[0].upper() + x[1:].lower())
print(df0.shape)
df0

# Files 

f = '../data_dump/P38_glut.h5ad'
adata = ad.read(f)
adata

# save specific genes

cgenes = np.intersect1d(adata.var.index.values, df0['Gene'].values)
print([g for g in df0['Gene'].values if g not in cgenes])
cgenes.shape

df0 = df0[df0['Gene'].isin(cgenes)]
df0

# f = '../data_dump/candidate_genes_vincent_0503_v2.csv'
# df0.to_csv(f, header=True, index=False)

f = '../data_dump/candidate_genes_vincent_0503_v2.csv'
df0 = pd.read_csv(f)
df0

%%time
adata = adata[:,cgenes]
adata

def rename_clsts(x):
    thedict = {
        'L2/3_A': 'L2/3_B',
        'L2/3_B': 'L2/3_A',
    }
    if x in thedict.keys():
        return thedict[x]
    else:
        return x

%%time
adata = adata[adata.obs['cluster'].str.contains(r'^L2|L4')]
adata.obs['cluster'] = adata.obs['cluster'].apply(rename_clsts) 
adata

adata.write('/bigstore/GeneralStorage/fangming/projects/visctx/data_dump/P38_glut_genes_v3.h5ad')
