==> analysis_beginning/00.check_data.ipynb <==
import pandas as pd
import numpy as np
import anndata as ad

# Metadata

f = '../data/candidate_genes_vincent_0316.xlsx'
df0 = pd.read_excel(f, sheet_name=0)
df1 = pd.read_excel(f, sheet_name=1)

print(df0.shape, df1.shape)

df0

df1

# Files 

f = '../data/L23_VD.h5ad'
adata1 = ad.read(f, backed='r+')
adata1

f = '../data/P38_glut.h5ad'
adata2 = ad.read(f, backed='r+')
adata2

f = '../data/P38_glut_dr.h5ad'
adata3 = ad.read(f, backed='r+')
adata3

# Next
- all data 
- candidate genes: type A,B,C equal? best? can they classify A,B,C?
- MERFISH data?

adata2.obs

adata1.obs

adata1.obs['sample'].value_counts().sort_index()

adata2.obs['sample'].value_counts().sort_index()

adata2.obs['cluster'].value_counts().sort_index()

adata2.obs['Subclass'].value_counts().sort_index()

# get cluster mean

adata2

adata2.X

adata2.var

%%time
genes = [
    'Cdh13', 
    'Trpc6',
    'Chrm2',
    'Gria3',
    'Whrn',
    'Baz1a',
    'Adamts2',
    'Kcna1',
    'Igfn1',
    'Epha3',
]
adata2[:,genes].write('/bigstore/GeneralStorage/fangming/visctx/P38_glut_10genes.h5ad')

adata3[:,genes].write('/bigstore/GeneralStorage/fangming/visctx/P38_glut_dr_10genes.h5ad')
