==> analysis_beginning/21.explore_fang2022_data.ipynb <==
import numpy as np
import pandas as pd
import os
import glob

pth = '/bigstore/binfo/mouse/Brain/FISH/MERFISH_Fang2022_Science_CTX'
!ls $pth

files = np.sort(glob.glob("mouse*", dir_fd=pth))
files

f = os.path.join(pth, "mouse1.AUD_TEA_VIS.242.unexpand.features.csv")
meta = pd.read_csv(f)
meta

f = os.path.join(pth, "mouse1.AUD_TEA_VIS.242.unexpand.genes.csv")
genes = pd.read_csv(f)
genes

f = os.path.join(pth, "mouse1.AUD_TEA_VIS.242.unexpand.matrix.csv")
mtx = pd.read_csv(f)
mtx
