==> analysis_beginning/22.explore_raw_counts_scrna-largescale-Copy1-Oct24.ipynb <==
import numpy as np
import pandas as pd
from scipy.io import mmread
import os
import glob
import anndata
import tqdm
import zarr

- merge marker genes (types ABC)
- feed into NN design

rootpth = '/bigstore/GeneralStorage/fangming/projects/visctx/data_dump/counts/scRNAseq_Cheng2022_Cell_v1ctx/'
dirs = np.sort(glob.glob("*", root_dir=rootpth))

dirs

[d for d in dirs if d.startswith("P28")]

newdirs = [d for d in dirs if d.startswith("P28_nr")]
newdirs

# pth = '/bigstore/GeneralStorage/fangming/projects/visctx/data_dump/counts/scRNAseq_Cheng2022_Cell_v1ctx/P38_nr_1_a/filtered_feature_bc_matrix'
# !ls $pth

for i, _dir in tqdm.tqdm(enumerate(newdirs)):
    pth = os.path.join(rootpth, _dir, 'filtered_feature_bc_matrix')
    output = os.path.join(pth, 'counts.h5ad')
    print(pth)
    print(i, output)
    
    # check
    if os.path.isfile(output):
        print(_dir + " skipped...")
        continue
    
    # fnames 
    fc = os.path.join(pth, 'barcodes.tsv.gz')
    fg = os.path.join(pth, 'features.tsv.gz')
    fm = os.path.join(pth, 'matrix.mtx')
    fmgz = os.path.join(pth, 'matrix.mtx.gz')
    
    if not os.path.isfile(fc): 
        print(_dir + " failed...")
        continue
        
    if not os.path.isfile(fg): 
        print(_dir + " failed...")
        continue
        
    if not (os.path.isfile(fm) or os.path.isfile(fmgz)):
        print(_dir + " failed...")
        continue

    # load
    cells = pd.read_csv(fc, sep='\t', header=None)[0].values
    genes = pd.read_csv(fg, sep='\t', header=None).rename(columns={0: 'id', 1: 'name'})[['id', 'name']]
    if os.path.isfile(fm):
        mtx = mmread(fm)
    elif os.path.isfile(fmgz):
        mtx = mmread(fmgz)
    print(cells.shape, genes.shape, mtx.shape)
    
    # check
    if len(cells) != mtx.shape[1]:
        print(_dir + " failed...")
        continue
        
        
    # adata 
    adata = anndata.AnnData(
        X=mtx.T.tocsc(), 
        obs=pd.DataFrame(index=cells), 
        var=genes,
    )
    adata.write(output)
