==> analysis_beginning/24.reformat_into_zarr.ipynb <==
import numpy as np
import pandas as pd
from scipy.io import mmread
import os
import glob
import anndata
import tqdm
import zarr
from sklearn.decomposition import PCA
import matplotlib.pyplot as plt
import seaborn as sns
import scanpy as sp
import zarr

from dredFISH.Utils import basicu
from dredFISH.Utils import powerplots

- merge marker genes (types ABC)
- feed into NN design

pth = '/bigstore/GeneralStorage/fangming/projects/visctx/data_dump/counts/scRNAseq_Cheng2022_Cell_v1ctx/P38_nr_1_a/filtered_feature_bc_matrix'
f = os.path.join(pth, 'counts.h5ad')
adata = anndata.read(f)
adata.obs.index = np.char.add(adata.obs.index.values.astype(str), "-P38_1a-P38")
adata

cov = np.asarray(adata.X.sum(axis=1)).reshape(-1,)
medcov = np.median(cov)

adata.obs['cov'] = cov
adata.obs['covfactor'] = cov/medcov

adata.obs

data_dir = '/bigstore/GeneralStorage/fangming/projects/visctx/data_dump/'

f = '../data_dump/candidate_genes_vincent_0503_v2.csv'
df0 = pd.read_csv(f)
df0 = df0.sort_values(['Subclass', 'Type/Gradient'])
df0 = df0[df0['Gene']!='Neurod1']
df0 = df0.set_index('Gene')
df0

df0['Type/Gradient'].unique(), df0['Subclass'].unique()

genes = df0.index.values
allen_genes = df0[~df0['Allen marker'].isnull()].index.values
allen_genes.shape

f = os.path.join(data_dir, 'genes_lasso_all83_May13.csv')
g1 = pd.read_csv(f, index_col=0)
print(g1.shape)

f = os.path.join(data_dir, 'genes_lasso_nonallen_May13.csv')
g2 = pd.read_csv(f, index_col=0)
print(g2.shape)

%%time
adatasub = anndata.read('../data_dump/P38_glut_genes_v3.h5ad')
adatasub

adatasub.obs['sample'].unique()

tmp = adata[adatasub[adatasub.obs['sample']=='P38_1a'].obs.index].copy()
unq, inv, cnt = np.unique(tmp.var['name'].values, return_counts=True, return_inverse=True)
tmp = tmp[:, cnt[inv] == 1]
genes_idx = basicu.get_index_from_array(tmp.var['name'].values.astype(str), df0.index.values.astype(str))
tmp = tmp[:, genes_idx]
tmp.var = tmp.var.set_index('name')
tmp

mat = np.asarray(tmp.X.todense())
matnorm = mat/tmp.obs['covfactor'].values.reshape(-1,1)
mat.shape, matnorm.shape

tmp.layers['normed'] = matnorm
tmp.obs = tmp.obs.join(adatasub.obs)
tmp.X = np.asarray(tmp.X.todense())

tmp

tmp.obs

tmp.X

tmp.layers['normed']

tmp

dat_dir = '/bigstore/GeneralStorage/fangming/projects/dredfish/data/'
output = os.path.join(dat_dir, 'rna', 'scrna_visctx_P38_test.zarr')
print(output)

%%time
chunksize = 10
cells = adata.obs.index.values
genes = adata.var.index.values
ncells = len(cells)
ngenes = len(genes)

# create a zarr file (dir)
z1 = zarr.open(
          output, 
          mode='w', 
    )
z1

%%time
# create count matrix
z1.create_dataset('counts', 
                  shape=(ncells, ngenes), 
                  chunks=(chunksize, None), 
                  dtype='i4',
                 )
z1['counts'][:] = adata.X

# create count matrix
z1.create_dataset('normed', 
                  shape=(ncells, ngenes), 
                  chunks=(chunksize, None), 
                  dtype='i4',
                 )
z1['normed'][:] = adata.layers['normed']

# create labels
z1.create_dataset(
    'l1_code',
    shape=(ncells,), 
    chunks=(chunksize,), 
    dtype='i4',
)
a = adata.obs['class_label']
z1['l1_code'] = a.cat.codes.values
z1['l1_cat'] = a.cat.categories.values.astype(str)

z1.create_dataset(
    'l2_code',
    shape=(ncells,), 
    chunks=(chunksize,), 
    dtype='i4',
)
a = adata.obs['neighborhood_label']
z1['l2_code'] = a.cat.codes.values
z1['l2_cat'] = a.cat.categories.values.astype(str)

z1.create_dataset(
    'l3_code',
    shape=(ncells,), 
    chunks=(chunksize,), 
    dtype='i4',
)
a = adata.obs['subclass_label']
z1['l3_code'] = a.cat.codes.values
z1['l3_cat'] = a.cat.categories.values.astype(str)

z1.create_dataset(
    'l5_code',
    shape=(ncells,), 
    chunks=(chunksize,), 
    dtype='i4',
)
a = adata.obs['cluster_label']
z1['l5_code'] = a.cat.codes.values
z1['l5_cat'] = a.cat.categories.values.astype(str)

%%time
y = z1['counts'].oindex[975,:]
y, y.shape

%%time
testidx = np.random.choice(ncells, 100)

%%time
z1['counts'].oindex[testidx,:].shape

%%time
z1['counts'].oindex[testidx[0],:].shape

%%time
z1['l3_code'].oindex[testidx[0]].shape

# split train and test

def split_train_test(zarr_file, keys_copy, keys_split, frac=0.9, random_seed=None):
    """randomly select frac vs 1-frac samples into training and test (validation) set.
    Save them as separate zarr files
    """
    assert frac <= 1 and frac >= 0
    # the original zarr file
    z = zarr.open(zarr_file, 'r')
    size = len(z['counts'])
    
    path_train = zarr_file.replace('.zarr', '_train.zarr')
    path_test = zarr_file.replace('.zarr', '_test.zarr')
    print(f"{zarr_file} -> \n{path_train} and \n{path_test}\n")
    if random_seed: np.random.seed(random_seed)
    cond_train = np.random.rand(size) < frac
    ntrain = cond_train.sum()
    ntest = (~cond_train).sum()
    print(f"{size}, {ntrain} ({ntrain/size:.3f}), {ntest} ({ntest/size:.3f})")
    
    z_train = zarr.open(path_train, mode='w')
    z_test = zarr.open(path_test, mode='w')
    for key in keys_copy:
        z_train[key] = z[key]
        z_test[key] = z[key]

    for key in keys_split:
        # train
        if z[key].ndim == 1:
            chunksize = (10,)
            dat = z[key].oindex[cond_train]
        elif z[key].ndim == 2:
            chunksize = (10,None)
            dat = z[key].oindex[cond_train,:]
        else:
            raise ValueError('unimplemented')
            
        z_train.create_dataset(key, shape=dat.shape, chunks=chunksize)
        z_train[key][:] = dat
        
        # test
        if z[key].ndim == 1:
            chunksize = (10,)
            dat = z[key].oindex[~cond_train]
        elif z[key].ndim == 2:
            chunksize = (10,None)
            dat = z[key].oindex[~cond_train,:]
        else:
            raise ValueError('unimplemented')
            
        z_test.create_dataset(key, shape=dat.shape, chunks=chunksize)
        z_test[key][:] = dat
        
    return 

z = zarr.open(output, mode='r')
z, list(z.keys())

#
keys_copy = ['num_probe_limit', 'l1_cat', 'l2_cat', 'l3_cat', 'l5_cat']
keys_split = [key for key in z.keys() if key not in keys_copy]
keys_copy, keys_split

split_train_test(output, keys_copy, keys_split, frac=0.9)

# test if we can load train and val as well as the original

f_org = '/bigstore/GeneralStorage/fangming/projects/dredfish/data/rna/scrna_ss_ctxhippo_a_exon_count_matrix_v3.zarr'
f_trn = '/bigstore/GeneralStorage/fangming/projects/dredfish/data/rna/scrna_ss_ctxhippo_a_exon_count_matrix_v3_train.zarr'
f_tst = '/bigstore/GeneralStorage/fangming/projects/dredfish/data/rna/scrna_ss_ctxhippo_a_exon_count_matrix_v3_test.zarr'


z_trn = zarr.open(f_trn, mode='r')
z_tst = zarr.open(f_tst, mode='r')
z_trn, list(z_trn.keys()), z_tst, list(z_tst.keys())

z_trn['num_probe_limit'][:]

%%time
y = z_trn['counts'].oindex[975,:]
y, y.shape

%%time
y = z_tst['counts'].oindex[975,:]
y, y.shape

# fix num probe limit (change r to a for read mode)

z_org = zarr.open(f_org, mode='r')
z_trn = zarr.open(f_trn, mode='r')
z_tst = zarr.open(f_tst, mode='r')
z_trn, list(z_trn.keys()), z_tst, list(z_tst.keys())

# z_org['num_probe_limit'][:] = 100
# z_trn['num_probe_limit'][:] = 100
# z_tst['num_probe_limit'][:] = 100

z_org['num_probe_limit'][:], z_trn['num_probe_limit'][:], z_tst['num_probe_limit'][:]

import torch

torch.tensor(0).item()
