==> analysis_beginning/27.check_v1-scrna-data-merge-P14nr-Oct26.ipynb <==
import numpy as np
import pandas as pd
from scipy.io import mmread
import os
import glob
import anndata
import tqdm
import zarr
from sklearn.decomposition import PCA
import matplotlib.pyplot as plt
import seaborn as sns

from dredFISH.Utils import basicu
from dredFISH.Utils import powerplots

%%time
pths = [
    "../data_dump/counts/scRNAseq_Cheng2022_Cell_v1ctx/processed/P14_glut.h5ad",
    "../data_dump/counts/scRNAseq_Cheng2022_Cell_v1ctx/processed/P14_gaba.h5ad",
    "../data_dump/counts/scRNAseq_Cheng2022_Cell_v1ctx/processed/P14_non.h5ad",
]
meta = []
for pth in pths:
    meta_ = anndata.read(pth, backed='r').obs
    meta.append(meta_)
    print(meta_.shape)
meta = pd.concat(meta, join='inner')
newnames = ['-'.join(i.split('-')[:4]) for i in meta.index.values]
meta.index = newnames
print(meta.shape)
meta = meta[~meta.index.duplicated(keep='first')]
print(meta.shape)
meta

suffixes = ['-'.join(i.split('-')[2:]) for i in meta.index.values]
np.unique(suffixes)

meta['sample'].unique()

ddir = '/bigstore/GeneralStorage/fangming/projects/visctx/data_dump/counts/'
pths = [
    os.path.join(ddir, "scRNAseq_Cheng2022_Cell_v1ctx/P14_nr_1_a/filtered_feature_bc_matrix/counts.h5ad"),
    os.path.join(ddir, "scRNAseq_Cheng2022_Cell_v1ctx/P14_nr_1_b/filtered_feature_bc_matrix/counts.h5ad"),
    os.path.join(ddir, "scRNAseq_Cheng2022_Cell_v1ctx/P14_nr_2_a/filtered_feature_bc_matrix/counts.h5ad"),
    os.path.join(ddir, "scRNAseq_Cheng2022_Cell_v1ctx/P14_nr_2_b/filtered_feature_bc_matrix/counts.h5ad"),
]
renames = [
    "-P14_1a-P14",
    "-P14_1b-P14",
    "-P14_2a-P14",
    "-P14_2b-P14",
]

adata = []
for f, rn in zip(pths, renames):
    adata_ = anndata.read(f)
    adata_.obs.index = np.char.add(adata_.obs.index.values.astype(str), rn)
    print(adata_.shape)
    adata.append(adata_)
    
adata = anndata.concat(adata, merge='same')
print(adata.shape)
adata

np.unique(adata.obs.index).shape

adata_.obs

cov = np.asarray(adata.X.sum(axis=1)).reshape(-1,)
medcov = np.median(cov)
adata.obs['cov'] = cov
adata.obs['covfactor'] = cov/medcov

tmp = adata[meta.index].copy()
unq, inv, cnt = np.unique(tmp.var['name'].values, return_counts=True, return_inverse=True)
tmp = tmp[:, cnt[inv] == 1]
tmp.var = tmp.var.set_index('name')
tmp.obs = tmp.obs.join(meta)
# tmp

tmp.var

tmp.obs

pth_res = os.path.join(ddir, "P14_nr_allcells_Oct26.h5ad")
tmp.write(pth_res)

b = anndata.read(pth_res, backed='r')
b

# subselect L234 only

tmp.shape, meta.shape

np.unique(meta.Type)

selected_types = ['L2/3_AB', 'L2/3_BC', 'L4_A', 'L4_B', 'L4_C']

tmp_sub = tmp[meta[meta['Type'].isin(selected_types)].index,:]
tmp_sub

pth_res2 = os.path.join(ddir, "P14_nr_L234_Oct26.h5ad")
tmp_sub.write(pth_res2)

b = anndata.read(pth_res2, backed='r')
b
