==> analysis_beginning/44.parse_gtf.ipynb <==
import pandas as pd

f = '../data_dump/gencode.vM32.annotation.gtf.gz'
df = pd.read_csv(f, sep="\t", skiprows=5, header=None)
df

dfg = df[df[2] == 'gene'].copy()
dfg 

dfg['gene_id']   = dfg[8].apply(lambda x: x.split(';')[0].split(' ')[1].strip('\"')) 
dfg['gene_name'] = dfg[8].apply(lambda x: x.split(';')[2].split(' ')[2].strip('\"'))
dfg

dfgbed = dfg[[0, 3, 4, 6, 'gene_id', 'gene_name',]].copy()
dfgbed['gid_short'] = dfgbed['gene_id'].apply(lambda x: x.split('.')[0])
dfgbed

fout = '../data_dump/gencode.vM32.allgenes.bed'
dfgbed.to_csv(fout, sep="\t", header=False, index=False)

!head $fout
