==> analysis_beginning/51.reproc_geo_data.ipynb <==
import os
import numpy as np
import pandas as pd
import anndata

f = "../data/cheng21_cell_scrna/counts/GSE190940_P8NR.csv.gz"
df = pd.read_csv(f, nrows=10)
df


files = [
    "GSE190940_P8NR.csv.gz",
    "GSE190940_P14NR.csv.gz",
    "GSE190940_P17NR.csv.gz",
    "GSE190940_P21NR.csv.gz",
    "GSE190940_P28NR.csv.gz",
    "GSE190940_P38NR.csv.gz",
    
    "GSE190940_P28DR.csv.gz",
    "GSE190940_P38DR.csv.gz",
    
    "GSE190940_P28DL.csv.gz",
] 

genes = []
ddir = '../data/cheng21_cell_scrna/counts/'

for f in files:
    path = os.path.join(ddir, f)
    df = pd.read_csv(path, nrows=5)
    _genes = df.columns[1:].values
    if len(genes) > 0:
        assert np.all(genes == _genes)
    else:
        genes = _genes
        
    print(df.shape)

genes.shape

unq, cnts = np.unique(genes, return_counts=True)
unq.shape, (cnts>1).sum()

"Bcl11b" in genes, "Magi1" in genes

f = "../data/cheng21_cell_scrna/processed/P8_glut.h5ad"
adata = anndata.read(f, backed='r')
adata

"Magi1" in adata.var.index.values

adata.var

adata.obs

files = [
    'P8_glut.h5ad',
    'P8_gaba.h5ad',
    'P8_non.h5ad',
    
    'P14_glut.h5ad',  
    'P14_gaba.h5ad',
    'P14_non.h5ad',  
    
    'P17_glut.h5ad', 
    'P17_gaba.h5ad',  
    'P17_non.h5ad',   
    
    'P21_glut.h5ad',  
    'P21_gaba.h5ad',  
    'P21_non.h5ad',     
    
    'P28_glut.h5ad',     
    'P28_gaba.h5ad',     
    'P28_non.h5ad',      
    
    'P38_glut.h5ad',    
    'P38_gaba.h5ad',    
    'P38_non.h5ad',   
    
    'P28_glut_dl.h5ad',  
    'P28dl_gaba.h5ad',  
    'P28_non_DL.h5ad',   
    
    'P28_glut_dr.h5ad', 
    'P28dr_gaba.h5ad',
    'P28_non_DR.h5ad',
    
    'P38_glut_dr.h5ad',  
    'P38dr_gaba.h5ad',
    'P38_non_DR.h5ad',   
    ]
ddir = "../data/cheng21_cell_scrna/processed"

for f in files:
    path = os.path.join(ddir, f)
    adata = anndata.read(path, backed='r')
    print(f,adata.shape)
