==> analysis_beginning/52.reproc_geo_data.ipynb <==
import os
import numpy as np
import pandas as pd
from scipy import sparse
import anndata
from scroutines import basicu

ddir1 = '../data/cheng21_cell_scrna/counts/'
files1 = [
    # "GSE190940_P8NR.csv.gz",
    # "GSE190940_P14NR.csv.gz",
    # "GSE190940_P17NR.csv.gz",
    # "GSE190940_P21NR.csv.gz",
    # "GSE190940_P28NR.csv.gz",
    # "GSE190940_P38NR.csv.gz",
    
    "GSE190940_P28DR.csv.gz",
    "GSE190940_P38DR.csv.gz",
    
    "GSE190940_P28DL.csv.gz",
] 

ddir2 = "../data/cheng21_cell_scrna/processed"
files2 = [
#     'P8_glut.h5ad',
#     'P8_gaba.h5ad',
#     'P8_non.h5ad',
    
#     'P14_glut.h5ad',  
#     'P14_gaba.h5ad',
#     'P14_non.h5ad',  
    
#     'P17_glut.h5ad', 
#     'P17_gaba.h5ad',  
#     'P17_non.h5ad',   
    
#     'P21_glut.h5ad',  
#     'P21_gaba.h5ad',  
#     'P21_non.h5ad',     
    
#     'P28_glut.h5ad',     
#     'P28_gaba.h5ad',     
#     'P28_non.h5ad',      
    
#     'P38_glut.h5ad',    
#     'P38_gaba.h5ad',    
#     'P38_non.h5ad',   
    
    'P28DR_glut.h5ad', 
    'P28DR_gaba.h5ad',
    'P28DR_non.h5ad',
    
    'P38DR_glut.h5ad',  
    'P38DR_gaba.h5ad',
    'P38DR_non.h5ad',   
    
    'P28DL_glut.h5ad',  
    'P28DL_gaba.h5ad',  
    'P28DL_non.h5ad',   
    ]

ddir3 = '../data/cheng21_cell_scrna/organized'

# for i, f in enumerate(files1[:6]):
#     sample = f[len('GSE190940_'):-len('.csv.gz')]
#     f1, f2, f3 = files2[3*i:3*i+3]
#     print(sample, f1, f2, f3)
    
#     path = os.path.join(ddir1, f)
#     df = pd.read_csv(path, index_col=0) #, nrows=5)
    
#     path = os.path.join(ddir2, f1)
#     adata1 = anndata.read(path, backed='r')
    
#     path = os.path.join(ddir2, f2)
#     adata2 = anndata.read(path, backed='r')
    
#     path = os.path.join(ddir2, f3)
#     adata3 = anndata.read(path, backed='r')
    
#     meta = pd.concat([adata1.obs, adata2.obs, adata3.obs])
    
#     # trim cell names to make it consistent
#     cellnames = ["-".join(i.split('-')[:4]) for i in meta.index.values]
#     meta.index = cellnames
    
#     # select only the cells in meta; assume all cells in meta are there in the counts
#     cellidx = basicu.get_index_from_array(df.index.values, cellnames) 
#     print(df.shape, meta.shape, cellidx.shape)
    
#     # assert np.all(cellidx != -1)
#     meta = meta[cellidx!=-1]
#     cellidx = cellidx[cellidx!=-1]
#     print(df.shape, meta.shape, cellidx.shape)
    
    
#     adata = anndata.AnnData(
#         X=sparse.csc_matrix(df.values[cellidx]),
#         obs=meta,
#         var=pd.DataFrame(index=df.columns.values),
#     )
    
#     fout = os.path.join(ddir3, f"{sample}.h5ad")
#     print(fout)
#     adata.write(fout)
#     # break

for i, f in enumerate(files1):
    sample = f[len('GSE190940_'):-len('.csv.gz')]
    f1, f2, f3 = files2[3*i:3*i+3]
    print(sample, f1, f2, f3)
    
    path = os.path.join(ddir1, f)
    df = pd.read_csv(path, index_col=0) #, nrows=5)
    
    path = os.path.join(ddir2, f1)
    adata1 = anndata.read(path, backed='r')
    
    path = os.path.join(ddir2, f2)
    adata2 = anndata.read(path, backed='r')
    
    path = os.path.join(ddir2, f3)
    adata3 = anndata.read(path, backed='r')
    
    meta = pd.concat([adata1.obs, adata2.obs, adata3.obs])
    
    # break
    # trim cell names to make it consistent
    cellnames = ["-".join(i.split('-')[:3]) for i in meta.index.values]
    meta.index = cellnames
    
    # select only the cells in meta; assume all cells in meta are there in the counts
    cellidx = basicu.get_index_from_array(df.index.values, cellnames) 
    print(df.shape, meta.shape, cellidx.shape)
    
    # assert np.all(cellidx != -1)
    meta = meta[cellidx!=-1]
    cellidx = cellidx[cellidx!=-1]
    print(df.shape, meta.shape, cellidx.shape)
    
    adata = anndata.AnnData(
        X=sparse.csc_matrix(df.values[cellidx]),
        obs=meta,
        var=pd.DataFrame(index=df.columns.values),
    )
    
    fout = os.path.join(ddir3, f"{sample}.h5ad")
    print(fout)
    adata.write(fout)
    # break

print('a')
