==> analysis_beginning/53.fix_dupcells_in_org_data.ipynb <==
import os
import numpy as np
import pandas as pd
from scipy import sparse
import anndata
from scroutines import basicu

ddir = '../data/cheng21_cell_scrna/organized/'
files = [
    # 'P8NR.h5ad',
    # 'P14NR.h5ad',
    # 'P17NR.h5ad',
    # 'P21NR.h5ad',
    # 'P28NR.h5ad',
    # 'P38NR.h5ad',
    
    'P28DR.h5ad',
    'P38DR.h5ad',
    
    # 'P28DL.h5ad',
] 

for f in files:
    path = os.path.join(ddir, f)
    adata = anndata.read(path, backed='r')
    # break

# first to fix

f = files[0]
print(f)
path = os.path.join(ddir, f)
adata = anndata.read(path) #, backed='r')
adata

cellnames = adata.obs.index.values
unq, cnts = np.unique(cellnames, return_counts=True)
unq[cnts>1]

adata.obs_names_make_unique()

adata.obs.filter(regex='^GTGTTCCTCTCAGGCG-1-P28_dr_1b', axis=0)

adata.obs.filter(regex='-1$', axis=0)

adata_trim = adata[unq[cnts==1]]

adata_trim.write(path.replace(".h5ad", ".u.h5ad"))

test = anndata.read(path.replace(".h5ad", ".u.h5ad"))
test 

test.obs.filter(regex='-1$', axis=0)

# second to fix

f = files[1]
print(f)
path = os.path.join(ddir, f)
adata = anndata.read(path, backed='r')
adata

cellnames = adata.obs.index.values
unq, cnts = np.unique(cellnames, return_counts=True)
unq[cnts>1]

adata.obs.loc[['GTGTTCCTCTCAGGCG-1-P38_dr_1b', 'TCCTAATTCACATTGG-1-P38_dr_1b']]

adata.obs_names_make_unique()

adata.obs.filter(regex='-1$', axis=0)

adata_trim = adata[unq[cnts==1]]
adata_trim

adata_trim.write(path.replace(".h5ad", ".u.h5ad"))

test = anndata.read(path.replace(".h5ad", ".u.h5ad"))
test 

test.obs.filter(regex='-1$', axis=0)
