==> analysis_beginning/58.clean_up_all_MERFISH_genes.ipynb <==
import itertools
import numpy as np
import pandas as pd
from scipy.stats import ttest_ind
import anndata

import matplotlib.pyplot as plt
from statsmodels.stats.multitest import multipletests
import seaborn as sns

from scroutines import basicu
# from scroutines import powerplots

sns.set_context('talk')

# base line

f = "../data/cheng21_cell_scrna/organized/P28NR.h5ad"

adata = anndata.read(f, backed='r')
genes = adata.var.index.values
genes


f = "../results/MERFISH_gene_panel_working_Feb28.csv"
df = pd.read_csv(f)

cnddts = df['gene_name_data'].values
unq, cnts = np.unique(cnddts, return_counts=True)
print(len(cnddts), unq.shape, unq[cnts>1], [g for g in cnddts if g not in genes])

df

df.groupby(['why included', 'source']).size().to_frame('number')

df.groupby(['why included']).size().to_frame('number')

df.groupby(['source']).size().to_frame('number')

cond1 = df['why included'].str.contains("All cell types")
cond2 = df['source'].str.contains("Chen22_biorxiv_Zador")
print(df[cond1].shape)
print(df[cond2].shape)
print(df[cond1 & cond2].shape)

# our analysis genes

f = '../results/dynamic_genes_230228.csv'
dfu1 = pd.read_csv(f)
dfu1

f = '../results/P28NRvsDR_DEGs_230228.csv'
dfu2 = pd.read_csv(f)
dfu2

# Merge 

dfu1r = dfu1[dfu1['subclass']=='L2/3'].copy()
assert len(dfu1r) == len(dfu1r['gene'].unique())
dfu1r['gene_name_data'] = dfu1r['gene']
dfu1r['gene_name_vizgen'] = dfu1r['gene']
dfu1r['why included'] = "L2/3 dynamic"
dfu1r['source'] = 'our analysis'
dfu1r

dfu2r = dfu2.groupby('gene', as_index=False).agg({'subclass': ",".join, 'reason': ",".join}).copy() 
dfu2r['gene_name_data'] = dfu2r['gene']
dfu2r['gene_name_vizgen'] = dfu2r['gene']
dfu2r['why included']   = 'NRvsDR_DEG'
dfu2r['source']   = 'our analysis'
dfu2r['Annot1'] = dfu2r['subclass']
dfu2r['Annot2'] = dfu2r['reason']
dfu2r

df

dfnew = df.copy()
dfnew

newidx1 = []
for idx, row in dfu1r.iterrows():
    if row['gene'] not in df['gene_name_data'].values:
        newidx1.append(idx)
len(newidx1)

newidx2 = []
for idx, row in dfu2r.iterrows():
    if ((row['gene'] not in df['gene_name_data'].values) and 
        (row['gene'] not in dfu1r['gene_name_data'].values)):
        newidx2.append(idx)
len(newidx2)

df_merged = pd.concat([df, 
                       dfu1r.loc[newidx1, ['gene_name_vizgen', 'gene_name_data', 'why included', 'source',]], 
                       dfu2r.loc[newidx2, ['gene_name_vizgen', 'gene_name_data', 'why included', 'source', 'Annot1', 'Annot2']], 
                       ], ignore_index=True)
df_merged

len(df_merged['gene_name_data'].unique()) == len(df_merged)

# add those gene annotations back
dfout = df_merged.copy()
print(dfout.shape)

g1 = dfu1r['gene'].values
g2 = dfu2r['gene'].values

cond = (~dfout['why included'].str.contains("L2/3 dynamic")) & dfout['gene_name_data'].isin(g1)
dfout.loc[cond,'why included'] = dfout.loc[cond, 'why included']+";L2/3 dynamic"

cond = (~dfout['source'].str.contains("our analysis")) & dfout['gene_name_data'].isin(g1)
dfout.loc[cond,'source'] = dfout.loc[cond, 'source']+";our analysis"

cond = (~dfout['why included'].str.contains("NRvsDR_DEG")) & dfout['gene_name_data'].isin(g2)
dfout.loc[cond,'why included'] = dfout.loc[cond, 'why included']+";NRvsDR_DEG"

cond = (~dfout['source'].str.contains("our analysis")) & dfout['gene_name_data'].isin(g2)
dfout.loc[cond,'source'] = dfout.loc[cond, 'source']+";our analysis"

dfout

# Save

fout = '../results/MERFISH_gene_panel_merged_Feb28.csv' 
dfout.to_csv(fout, header=True, index=False)

!head $fout

!wc -l $fout
