==> analysis_beginning/62.check_MERFISH_genes_Mar7.ipynb <==
import itertools
import numpy as np
import pandas as pd
from scipy.stats import ttest_ind
import anndata

import matplotlib.pyplot as plt
from statsmodels.stats.multitest import multipletests
import seaborn as sns

from scroutines import basicu
# from scroutines import powerplots

sns.set_context('talk')

# base line

f = "../data/cheng21_cell_scrna/organized/P28NR.h5ad"
adata = anndata.read(f, backed='r')
genes = adata.var.index.values
genes

f = "../results/MERFISH_gene_panel_Current_Mar7.csv"
df = pd.read_csv(f)

# remove Vincent genes
f = "../results/MERFISH_gene_panel_VX_edit_Mar7.csv"
df2 = pd.read_csv(f)
df = df[df['gene_name_data'].isin(df2['gene_name_data'])].copy()

cnddts = df['gene_name_data'].values
unq, cnts = np.unique(cnddts, return_counts=True)
print(len(cnddts), unq.shape, unq[cnts>1], [g for g in cnddts if g not in genes])

df

# annotate and add

f = "../data/cheng21_cell_scrna/res/L23-ABC-genes-n288-n286unq-annot.csv"
df3 = pd.read_csv(f)
df3['Annot4'] = 'L2/3 types'
df3

df3['gene'].isin(cnddts).sum(), (~df3['gene'].isin(cnddts)).sum()

g3_p  = df3[df3['ispicked']]['gene']
g3_up = df3[~df3['ispicked']]['gene']

# do not add g3_p
g0 = g3_p.values
print(np.all(g3_p.isin(cnddts)))
# add annot
g1 = g3_up[g3_up.isin(cnddts)].values
# add unique
g2 = g3_up[~g3_up.isin(cnddts)].values

print(g0.shape)
print(g1.shape)
print(g2.shape)
print(len(g0)+len(g1)+len(g2))

dfout = df.copy()
cond = dfout['gene_name_data'].isin(g1)
dfout.loc[cond,'why included'] = "L2/3 subtypes;"+dfout.loc[cond, 'why included']
dfout.loc[cond,'source']       = "Cheng22_Cell;" +dfout.loc[cond, 'source']

cond = dfout['gene_name_data'].isin(g2)
cond.sum()

dfnew = pd.DataFrame()
dfnew['gene_name_data'] = g2
dfnew['gene_name_vizgen'] = g2
dfnew['why included'] = 'L2/3 subtypes'
dfnew['source'] = 'Cheng22_Cell'
dfout = pd.concat([dfout, dfnew], ignore_index=True)


dfout = pd.merge(dfout, df3[['gene', 'Annot4', 'ispicked', 'P17on', 'P14', 'DR']].rename(columns={'gene':'gene_name_data'}), 
                 how='left', on='gene_name_data')
dfout

fout = "../results/MERFISH_gene_panel_Current_Mar7-v2.csv"
dfout.to_csv(fout, header=True, index=False) 

dfout['ispicked'].sum()

(dfout['Annot4']=='L2/3 types').sum()

# summarize

unq_reasons = np.unique(np.hstack(df['why included'].str.split(';')))
unq_reasons

breakdown = []
for reason in unq_reasons:
    num = df['why included'].str.contains(reason).sum()
    breakdown.append({'reason': reason, 'num': num})
breakdown = pd.DataFrame(breakdown)
breakdown.sort_values('num', ascending=False)

pd.set_option('display.max_rows', 100)
df.groupby(['why included', 'source']).size().to_frame('number')

df.groupby(['why included']).size().to_frame('number')

df.groupby(['why included']).size().sort_values(ascending=False).to_frame('number')

df.groupby(['source']).size().to_frame('number')

cond1 = df['why included'].str.contains("All cell types")
cond2 = df['source'].str.contains("Chen22_biorxiv_Zador")
print(df[cond1].shape)
print(df[cond2].shape)
print(df[cond1 & cond2].shape)

"March3" in df['gene_name_data'].values

# Save

# fout = '../results/MERFISH_gene_panel_merged_Feb28.csv' 
# dfout.to_csv(fout, header=True, index=False)

# !head $fout

# !wc -l $fout
